[NEW] Laid out file structure for components.
