--- server/server.js
@@ -1,8 +1,0 @@
-import express from "express";
-const port = '3000';
-// Setup static express server
-const app = express();
-app.use(express.static('build'));
-const listenPort = process.env.PORT || port;
-app.listen(listenPort);
-console.log(`Listening on port ${listenPort}`);
