+++ src/components/header/Header.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Header() {
+  return (
+    <div>
+      <h1>This is a header</h1>
+    </div>
+  );
+}
+
+export default Header;
