+++ src/components/index.tsx
@@ -1,0 +1,2 @@
+export * from "./header";
+export * from "./sidebar";
