+++ src/components/sidebar/Sidebar.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Sidebar() {
+  return (
+    <div>
+      <h1>This is a header</h1>
+    </div>
+  );
+}
+
+export default Sidebar;
