+++ src/components/sidebar/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Sidebar";
